scheduler:  Adjust port count migration to add a new attribute.
Changing the cardinality of an existing attribute in a migration won't
work, because we aren't ensuring that migrations are run only once -
so as soon as this change goes into effect, the original schema can
no longer be re-transacted and Cook won't start up.

New attribute is called "ports" rather than "port", which is a better
name for a single cardinality relationship indicating port count.