Remove the distinction of backfilled jobs (#241)
* Remove the distinction of backfilled jobs

The concept of backfilled jobs was to allow Cook to schedule out of
order (specifically, skip the head of the queue) to oppurtunistically
improve utilization. However, this can lead to cases where the job at
the head of the queue is starved and not able to get scheduled. When
this occurs, we rely on the rebalancer to free up space such that the
head of the queue can run. Instances were explicitly marked as
backfilled when Cook scheduled out of order and were then upgraded out
of backfilled status once the jobs ahead of them had been scheduled.
Instances that were marked as backfilled were treated as the first
jobs to be preempted.

This caused a variety of problems:
1. Preempting backfilled jobs first caused unexpected behavior when a
backfilled, high priority job was preempted instead of a low priority
job
2. Upgrading jobs from backfilled to fully processed is error prone and
causes issues the ordering of the queue. Specifically, jobs that were
backfilled would remain in the queue (generally high in the queue) and
would be removed when the jobs ahead of them were scheduled. However, in
the event that the backfilled job failed, it would "jump the queue"
which means that jobs that fail very quickly that get backfilled can
wreck havoc on the utilization of the cluster

We are choosing to remove the backfill distinction (while still allowing
cook to schedule out of order) because it of the problems above and
because the rebalancer will still be able to preempt jobs to make room
for starved jobs, it just may not be the backfilled jobs.

* Add doc string deprecated

* Factor out common logic

* fix