scheduler: On /rawscheduler, persist entities in post! rather than processable?
The code to save the jobs is unchanged.  However, this will mean that
"unforeseen" errors in the cook server will now correctly result in a 500
response code than a 422 ("unprocessable entity").

No validation was actually occurring in processable?, so until this change,
the only way a 422 would be returned is if an unforeseen error occurred.  500
is the appropriate response in this scenario.