Mf/409 on dupe job UUID (#260)
* scheduler: Extract test helper method to retrieve json body from response

* scheduler: On /rawscheduler, persist entities in post! rather than processable?

The code to save the jobs is unchanged.  However, this will mean that
"unforeseen" errors in the cook server will now correctly result in a 500
response code than a 422 ("unprocessable entity").

No validation was actually occurring in processable?, so until this change,
the only way a 422 would be returned is if an unforeseen error occurred.  500
is the appropriate response in this scenario.

* scheduler: Return a 409 (conflict) on /rawscheduler for duplicate job UUIDs.

* scheduler: docs: Bring docs about /rawscheduler return codes up to date.