Retract larger groups when retracting uncommitted jobs.
Each transaction is going to have a cost, so lets fewer transactions.

Jobs inserted at the same time as part of an uncommitted group are likely
closely colocated in storage, so this is likely going to be a lot fewer
random reads than you might otherwise expect; an interval of the
datomic storage tree to be retracted.