Move some other handlers from a compute-cluster/kill to compute-cluster/kill-task-if-possible.
There are cases, in particular, lingering tasks, stragglers, or cancelled tasks where the task might
outlive the compute cluster it was member of. When this occurs, the looked up compute cluster
is nil and trying to kill via it would cause an NPE. Switch them to using this wrapper function.