Workaround bug in running with minimesos.
We do testing in minimesos, which is unmaintained and 2 years old.
Our cook-executor docker image is based on python. As of 6 days ago,
its been updated. It now requires a version of glibc that is too new
and unsupported in minimesos's images. Temporary fix to force a specific
python image.