Reconfigure how we run integration tests.
We want which integration tests to run to be driven externally, so that way we can bring up one instance of cook
and run both mesos and k8s integration test suites on it. (and even hybrid tests)

To do this, we have knobs that let us control what pools the suite needs, and we inform it whether its in
mesos mode or k8s mode.