Makes test_preemption_basic respect the default submit pool (#1936)
Changes proposed in this PR

    making test_preemption_basic use the default submit pool for testing, if it's set and the rebalancer is enabled on it
    deleting test_preemption_for_pools

Why are we making these changes?

test_preemption_for_pools was attempting to test preemption on every active pool, which was overkill and broken. test_preemption_basic was not respecting the default submit pool for testing, and instead using the default pool of the cluster under test, which was broken in cases where that pool didn't have the rebalancer enabled but some other pool in the cluster did.